# ARROW-1572: [C++] Implement "value counts" kernels for tabulating value frequencies