--- a/cpp/src/arrow/array/builder_primitive.h
+++ b/cpp/src/arrow/array/builder_primitive.h
@@ -92,9 +92,7 @@ class NumericBuilder : public ArrayBuilder {
     return ArrayBuilder::Resize(capacity);
   }
 
-  value_type operator[](int64_t index) const {
-    return GetValue(index);
-  }
+  value_type operator[](int64_t index) const { return GetValue(index); }
 
   value_type& operator[](int64_t index) {
     return reinterpret_cast<value_type*>(data_->mutable_data())[index];