… cpp/src/arrow/array-test.cc
… cpp/src/arrow/compute/kernels/boolean-test.cc
… cpp/src/arrow/compute/kernels/hash-test.cc
… cpp/src/arrow/compute/kernels/hash.h
… cpp/src/arrow/util/hashing.h