Picked up from: https://github.com/apache/arrow/pull/1970

I mostly reused the unit tests as is and modified the rest based on feedback on that PR and adapting to new code.  Some other changes I mae:
1.  Remove the HashKernel and getter from the public API.  I think we can add these back once we have a better idea on how we are doing stateful kernels (i.e. https://github.com/apache/arrow/pull/3407)
2. Add operator[] to NumericBuilder to modify previously added values (this might not be the right place, and I had a little bit of trouble figuring out how to integrate this into the existing TypedTest so the testing is a little bit weak).  This seemed better then using a vector to collect values.

If this approach looks OK for now.  I'm also going to open up a JIRA to try refactor the unittest for Hash kernels (and maybe the headers) since I think there might be a clearer more granular way of laying these out.

other things to discuss:
1.  Handling null values in Count/Unique (it looks like they are dropped today, should this configurable/turned on).
2.  Hashing edge cases for floating point numbers (just opened a JIRA on this).