@emkornfield 2019-02-12
@fsaintjacques thanks for the review.  My thoughts:

* May I propose to rename this to `Distribution`?
** I renamed to ValueCounts to match up with corresponding Pandas function (https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.value_counts.html), Distribution might also make sense.  Since naming is hard, I'd like to get a second opinion  before changing.
* We should also count Nulls.
** I think this should be configurable, similar to Pandas, (and affects Uniques) as well, since this is a bigger change, I'd like to handle under a different JIRA if that is OK.
* The current output format is 2 arrays. I think we should formalize this and return a single array of `List<Struct<Option<value>, count>>` explicitly which should allow returning the count for the null value.
** Agreed, I am working through making this happen and will update the PR when I complete it.