@emkornfield 2019-02-27
@fsaintjacques I think I fixed the windows build (will check in the morning).  I changed the output to be a Struct as suggested.  I will file follow-up JIRAs for the rest.