@fsaintjacques 2019-03-02
I think you need to rebase for the failing appveyor failure, but I'd say this is safe to merge. LGTM +1.