@emkornfield 2019-03-06
@wesm or @xhochy mind taking a look to merge.  I believe the build failure was spurious due to https://github.com/apache/arrow/pull/3821