@wesm 2019-03-06
Welp, bad news. The performance regression makes this patch unmergeable for now

master branch

```
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
BM_BuildDictionary/min_time:1.000                                       2337 us       2337 us        617   1.66998GB/s
BM_BuildStringDictionary/min_time:1.000                                 3947 us       3947 us        358   76.4939MB/s
BM_UniqueInt64NoNulls/16777216/50/min_time:1.000/real_time             30889 us      30888 us         45   4.04681GB/s
BM_UniqueInt64NoNulls/16777216/1024/min_time:1.000/real_time           28470 us      28469 us         48   4.39063GB/s
BM_UniqueInt64NoNulls/16777216/10240/min_time:1.000/real_time          45673 us      45672 us         31   2.73684GB/s
BM_UniqueInt64NoNulls/16777216/1048576/min_time:1.000/real_time       424376 us     424374 us          3   301.619MB/s
BM_UniqueInt64WithNulls/16777216/50/min_time:1.000/real_time           44474 us      44474 us         31   2.81063GB/s
BM_UniqueInt64WithNulls/16777216/1024/min_time:1.000/real_time         38880 us      38879 us         35   3.21505GB/s
BM_UniqueInt64WithNulls/16777216/10240/min_time:1.000/real_time        55346 us      55346 us         25   2.25852GB/s
BM_UniqueInt64WithNulls/16777216/1048576/min_time:1.000/real_time     393982 us     393977 us          4   324.888MB/s
BM_UniqueString10bytes/16777216/50/min_time:1.000/real_time           217200 us     217199 us          7   736.649MB/s
BM_UniqueString10bytes/16777216/1024/min_time:1.000/real_time         184280 us     184279 us          7   868.243MB/s
BM_UniqueString10bytes/16777216/10240/min_time:1.000/real_time        298878 us     298877 us          5   535.335MB/s
BM_UniqueString10bytes/16777216/1048576/min_time:1.000/real_time     1895167 us    1895146 us          1   84.4253MB/s
BM_UniqueString100bytes/16777216/50/min_time:1.000/real_time          324695 us     324692 us          4   4.81221GB/s
BM_UniqueString100bytes/16777216/1024/min_time:1.000/real_time        407262 us     407258 us          3    3.8366GB/s
BM_UniqueString100bytes/16777216/10240/min_time:1.000/real_time       579285 us     579277 us          2   2.69729GB/s
BM_UniqueString100bytes/16777216/1048576/min_time:1.000/real_time    2684019 us    2683992 us          1   596.121MB/s
BM_UniqueUInt8NoNulls/16777216/200/min_time:1.000/real_time             9999 us       9998 us        140   1.56273GB/s
BM_UniqueUInt8WithNulls/16777216/200/min_time:1.000/real_time          20517 us      20516 us         68   779.853MB/s
```

this PR

```
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
BM_BuildDictionary/min_time:1.000                                       2964 us       2964 us        486   1.31671GB/s
BM_BuildStringDictionary/min_time:1.000                                 4740 us       4740 us        296   63.6928MB/s
BM_UniqueInt64NoNulls/16777216/50/min_time:1.000/real_time             58138 us      58137 us         24   2.15007GB/s
BM_UniqueInt64NoNulls/16777216/1024/min_time:1.000/real_time           53214 us      53213 us         25   2.34903GB/s
BM_UniqueInt64NoNulls/16777216/10240/min_time:1.000/real_time          73082 us      73081 us         19   1.71041GB/s
BM_UniqueInt64NoNulls/16777216/1048576/min_time:1.000/real_time       518587 us     518583 us          3   246.824MB/s
BM_UniqueInt64WithNulls/16777216/50/min_time:1.000/real_time           79186 us      79185 us         18   1.57857GB/s
BM_UniqueInt64WithNulls/16777216/1024/min_time:1.000/real_time         69846 us      69845 us         20   1.78966GB/s
BM_UniqueInt64WithNulls/16777216/10240/min_time:1.000/real_time        86503 us      86502 us         16   1.44504GB/s
BM_UniqueInt64WithNulls/16777216/1048576/min_time:1.000/real_time     477864 us     477860 us          3   267.859MB/s
BM_UniqueString10bytes/16777216/50/min_time:1.000/real_time           265399 us     265396 us          5   602.865MB/s
BM_UniqueString10bytes/16777216/1024/min_time:1.000/real_time         231742 us     231739 us          6   690.422MB/s
BM_UniqueString10bytes/16777216/10240/min_time:1.000/real_time        350534 us     350530 us          4   456.447MB/s
BM_UniqueString10bytes/16777216/1048576/min_time:1.000/real_time     1891871 us    1891853 us          1   84.5723MB/s
BM_UniqueString100bytes/16777216/50/min_time:1.000/real_time          378638 us     378635 us          4   4.12664GB/s
BM_UniqueString100bytes/16777216/1024/min_time:1.000/real_time        461684 us     461680 us          3   3.38435GB/s
BM_UniqueString100bytes/16777216/10240/min_time:1.000/real_time       637809 us     637804 us          2   2.44979GB/s
BM_UniqueString100bytes/16777216/1048576/min_time:1.000/real_time    2997666 us    2997611 us          1   533.749MB/s
BM_UniqueUInt8NoNulls/16777216/200/min_time:1.000/real_time            39592 us      39592 us         35    404.12MB/s
BM_UniqueUInt8WithNulls/16777216/200/min_time:1.000/real_time          53375 us      53375 us         26   299.764MB/s
```