@wesm 2019-03-06
By the way, I think it's growing less and less responsible by the day for us to continue to merge patches to master without running automated perf regression tests. @kszucs what do you think the road is like to set up a bot that can do automated reporting on major perf changes on PRs?