@emkornfield 2019-03-08
@wesm thanks for review and running the benchmark, I should have done it myself.  I will try to have a new version with hopefully equivalent performances numbers by sometime next week.