@wesm 2019-03-14
I ran on a mobile Xeon 2.8 ghz with the cores pinned at 2.8ghz / no turbo boost and things looking OK here

```
before

```
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
BM_BuildDictionary/min_time:1.000                                       2950 us       2950 us        476   1.32273GB/s
BM_BuildStringDictionary/min_time:1.000                                 5941 us       5940 us        234   50.8258MB/s
BM_UniqueInt64NoNulls/16777216/50/min_time:1.000/real_time             46613 us      46614 us         30   2.68166GB/s
BM_UniqueInt64NoNulls/16777216/1024/min_time:1.000/real_time           44150 us      44142 us         27   2.83129GB/s
BM_UniqueInt64NoNulls/16777216/10240/min_time:1.000/real_time          65805 us      65805 us         21   1.89955GB/s
BM_UniqueInt64NoNulls/16777216/1048576/min_time:1.000/real_time       505750 us     505756 us          3    253.09MB/s
BM_UniqueInt64WithNulls/16777216/50/min_time:1.000/real_time           70311 us      70312 us         20   1.77782GB/s
BM_UniqueInt64WithNulls/16777216/1024/min_time:1.000/real_time         62900 us      62901 us         22   1.98729GB/s
BM_UniqueInt64WithNulls/16777216/10240/min_time:1.000/real_time        85752 us      85754 us         17   1.45769GB/s
BM_UniqueInt64WithNulls/16777216/1048576/min_time:1.000/real_time     486725 us     486734 us          3   262.982MB/s
BM_UniqueString10bytes/16777216/50/min_time:1.000/real_time           308199 us     308202 us          4   519.145MB/s
BM_UniqueString10bytes/16777216/1024/min_time:1.000/real_time         262568 us     262572 us          5   609.367MB/s
BM_UniqueString10bytes/16777216/10240/min_time:1.000/real_time        439183 us     439189 us          3   364.313MB/s
BM_UniqueString10bytes/16777216/1048576/min_time:1.000/real_time     2329791 us    2329828 us          1   68.6757MB/s
BM_UniqueString100bytes/16777216/50/min_time:1.000/real_time          693907 us     693920 us          2   2.25174GB/s
BM_UniqueString100bytes/16777216/1024/min_time:1.000/real_time        821354 us     821368 us          2   1.90235GB/s
BM_UniqueString100bytes/16777216/10240/min_time:1.000/real_time      1094868 us    1094886 us          1   1.42711GB/s
BM_UniqueString100bytes/16777216/1048576/min_time:1.000/real_time    3770885 us    3770950 us          1   424.304MB/s
BM_UniqueUInt8NoNulls/16777216/200/min_time:1.000/real_time            13690 us      13690 us        102   1.14134GB/s
BM_UniqueUInt8WithNulls/16777216/200/min_time:1.000/real_time          31576 us      31574 us         45   506.714MB/s
```

after

```
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
BM_BuildDictionary/min_time:1.000                                       2969 us       2969 us        472   1.31424GB/s
BM_BuildStringDictionary/min_time:1.000                                 5966 us       5966 us        235   50.6048MB/s
BM_UniqueInt64NoNulls/16777216/50/min_time:1.000/real_time             46763 us      46763 us         30   2.67303GB/s
BM_UniqueInt64NoNulls/16777216/1024/min_time:1.000/real_time           44425 us      44426 us         32   2.81373GB/s
BM_UniqueInt64NoNulls/16777216/10240/min_time:1.000/real_time          69868 us      69869 us         21    1.7891GB/s
BM_UniqueInt64NoNulls/16777216/1048576/min_time:1.000/real_time       511248 us     511255 us          3   250.368MB/s
BM_UniqueInt64WithNulls/16777216/50/min_time:1.000/real_time           70771 us      70770 us         20   1.76626GB/s
BM_UniqueInt64WithNulls/16777216/1024/min_time:1.000/real_time         63143 us      63143 us         22   1.97964GB/s
BM_UniqueInt64WithNulls/16777216/10240/min_time:1.000/real_time        83603 us      83604 us         17   1.49516GB/s
BM_UniqueInt64WithNulls/16777216/1048576/min_time:1.000/real_time     486510 us     486505 us          3   263.098MB/s
BM_UniqueString10bytes/16777216/50/min_time:1.000/real_time           305595 us     305600 us          5   523.569MB/s
BM_UniqueString10bytes/16777216/1024/min_time:1.000/real_time         251144 us     251149 us          6   637.084MB/s
BM_UniqueString10bytes/16777216/10240/min_time:1.000/real_time        428113 us     428118 us          3   373.733MB/s
BM_UniqueString10bytes/16777216/1048576/min_time:1.000/real_time     2338876 us    2338908 us          1   68.4089MB/s
BM_UniqueString100bytes/16777216/50/min_time:1.000/real_time          691089 us     691102 us          2   2.26093GB/s
BM_UniqueString100bytes/16777216/1024/min_time:1.000/real_time        803811 us     803825 us          2   1.94387GB/s
BM_UniqueString100bytes/16777216/10240/min_time:1.000/real_time      1103679 us    1103689 us          1   1.41572GB/s
BM_UniqueString100bytes/16777216/1048576/min_time:1.000/real_time    4039441 us    4039457 us          1   396.094MB/s
BM_UniqueUInt8NoNulls/16777216/200/min_time:1.000/real_time            13894 us      13894 us        101   1.12461GB/s
BM_UniqueUInt8WithNulls/16777216/200/min_time:1.000/real_time          30398 us      30398 us         46   526.348MB/s
```

merging. thanks @emkornfield!